+++ R/calc_het.R
@@ -8,6 +8,11 @@
 #'
 #' @param omit_x If TRUE, omit the X chromosome.
 #'
+#' @param cores Number of CPU cores to use, for parallel calculations.
+#' (If `0`, use [parallel::detectCores()].)
+#' Alternatively, this can be links to a set of cluster sockets, as
+#' produced by [parallel::makeCluster()].
+#'
 #' @return
 #' The result is a vector of estimated heterozygosities
 #'
@@ -23,12 +28,15 @@
 #'
 #' @export
 calc_het <-
-    function(probs, by=c("individual", "marker"), omit_x=TRUE)
+    function(probs, by=c("individual", "marker"), omit_x=TRUE, cores=1)
 {
     by <- match.arg(by)
     if(is.null(probs)) stop("probs is NULL")
     if(is.cross2(probs))
         stop('Input probs is a "cross2" object but should be genotype probabilities, as from calc_genoprob')
+
+    # set up cluster
+    cores <- setup_cluster(cores, quiet=TRUE)
 
     is_x_chr <- attr(probs, "is_x_chr")
     if(any(is_x_chr) && !all(is_x_chr) && omit_x) {
@@ -54,7 +62,7 @@ calc_het <-
         total_mar <- sum(dim(probs)[3,])
 
         # summarize each chromosome
-        result <- lapply(seq_len(n_chr), function(chr) apply(probs[[chr]][,het_col[[chr]],,drop=FALSE], 1, sum))
+        result <- cluster_lapply(cores, seq_len(n_chr), function(chr) apply(probs[[chr]][,het_col[[chr]],,drop=FALSE], 1, sum))
 
         if(length(result)>1) {
             for(i in seq_along(result)[-1])
@@ -64,6 +72,6 @@ calc_het <-
     }
 
     # else: by marker
-    unlist(lapply(seq_len(n_chr), function(chr) apply(probs[[chr]][,het_col[[chr]],,drop=FALSE], 3, sum)/
+    unlist(cluster_lapply(cores, seq_len(n_chr), function(chr) apply(probs[[chr]][,het_col[[chr]],,drop=FALSE], 3, sum)/
                                                           nrow(probs[[chr]])))
 }